+++ CHANGELOG.md
@@ -17,3 +17,10 @@
 - issue #1, allow arbitrary environment name, https://github.com/praized/raad/issues/1
 - issue #3, avoid calling stop multiple times,  https://github.com/praized/raad/issues/3
 - added Raad.stopped?
+
+# 0.5.0, 10-12-2011
+- issue #4, allow logger usage in service initialize method, https://github.com/praized/raad/issues/4
+- issue #7, Raad.env not correct in service.initialize, https://github.com/praized/raad/issues/7
+- refactored Runner initialize/run/start_service methods
+- custom options now uses options_parser class method in the service
+- added examples/custom_options.rb